Update monitored zfs docs
Fixes #56.

The monitored zfs docs currently specify that the zfs software should be
installed before adding the servers to IML. The problem is that if you
attempt to install zfs on the servers before the agent is deployed it
will not be able to find the packages it needs to install zfs. The docs
should be updated to reflect this.

Signed-off-by: Will Johnson <william.c1.johnson@gmail.com>