Fix location constraint operation in 2.4 upgrade docs (#133)
The location constraints in the upgrade docs are being set using "pcs
constraint location ...". This will set the constraint but the format of
the constraint will not match what IML is expecting. As a result,
operations on the resource issued by IML will fail becasue it will not
be able to match the resource name. The IML source shows that the
constraints are being added using cibadmin. The instructions should use
the same command such that the resource names match.

Signed-off-by: Will Johnson <wjohnson@whamcloud.com>