Update uninstall instructions for the agent side
Starting in IML 5.0.0 files have been added to the /etc/iml/ directory.
If this folder is not removed after uninstalling and then the server is
used again it will contain cached data that will prevent volumes from
being picked up by iml. This patch updates the "Remove Agent Software"
section to ensure that this directory is removed.

Signed-off-by: johnsonw <wjohnson@whamcloud.com>